+++ RELEASE_NOTES.md
@@ -5,3 +5,6 @@ v0.67.6
 - keytrans: Remove unused APIs
 
 - backups: Relax check on session switchover update message authors
+
+- Node: Don't error if an unauthenticated chat socket receives an empty list of
+  alerts from the server.